change name from jijmodeling-transpiler-quantum to qamomile
