+++ qamomile/core/__init__.py
@@ -1,5 +1,5 @@
-from jijmodeling_transpiler_quantum.core import ising_qubo as ising_qubo
-from jijmodeling_transpiler_quantum.core import qrac as qrac
+from qamomile.core import ising_qubo as ising_qubo
+from qamomile.core import qrac as qrac
 from .ising_qubo import qubo_to_ising, IsingModel
 from .qrac import greedy_graph_coloring
