+++ qamomile/qiskit/minimal_encoding/cost_function.py
