+++ qamomile/qiskit/minimal_encoding/to_minimal_encoding.py
