+++ tests/qiskit/test_qrao32.py
@@ -2,7 +2,7 @@ import numpy as np
 import qiskit as qk
 import qiskit.quantum_info as qk_ope
 
-from jijmodeling_transpiler_quantum.qiskit.qrao.qrao32 import (
+from qamomile.qiskit.qrao.qrao32 import (
     create_pauli_x_prime_term,
     create_pauli_y_prime_term,
     create_pauli_z_prime_term,
