+++ tests/qiskit/test_qrao_space_efficient.py
@@ -4,9 +4,9 @@ import numpy as np
 import qiskit as qk
 import qiskit.quantum_info as qk_ope
 
-from jijmodeling_transpiler_quantum.core.ising_qubo import IsingModel
-from jijmodeling_transpiler_quantum.qiskit.qrao.qrao31 import Pauli
-from jijmodeling_transpiler_quantum.qiskit.qrao.qrao_space_efficient import (
+from qamomile.core.ising_qubo import IsingModel
+from qamomile.qiskit.qrao.qrao31 import Pauli
+from qamomile.qiskit.qrao.qrao_space_efficient import (
     numbering_space_efficient_encode,
     qrac_space_efficient_encode_ising,
 )
