+++ tests/quri_parts/test_to_qrac_quri.py
@@ -3,7 +3,7 @@ import jijmodeling_transpiler as jmt
 import numpy as np
 import qiskit as qk
 
-from jijmodeling_transpiler_quantum.quri_parts import (
+from qamomile.quri_parts import (
     transpile_to_qrac21_hamiltonian,
     transpile_to_qrac31_hamiltonian,
     transpile_to_qrac32_hamiltonian,
